+++ app/src/main/java/com/example/fitnesapp/models/firebase/HealthLogItem.java
@@ -1,13 +1,13 @@
 package com.example.fitnesapp.models.firebase;
 
 public class HealthLogItem {
-    public int val;      // Значение пульса или кислорода
+    public double val;      // Значение пульса или кислорода
     public long time;    // Timestamp (время в миллисекундах)
 
     public HealthLogItem() {}
 
     // Конструктор для удобства добавления
-    public HealthLogItem(int val, long time) {
+    public HealthLogItem(double val, long time) {
         this.val = val;
         this.time = time;
     }
